# internal/api/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Awo Enterprise Resource Planning Multi-Tenant API","description":"Comprehensive ERP system with RBAC/ABAC, finance, HR, and organizational management","version":"1.0.0"},"host":"api.erp.example.com","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/openapi.json":{"get":{"tags":["openapi"],"summary":"spec openapi","description":"Get OpenAPI specification in JSON format","operationId":"openapi#spec","responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/OpenapiSpecResponseBody","required":["spec"]}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/OpenapiSpecInternalErrorResponseBody"}}},"schemes":["https"]}},"/swagger-ui":{"get":{"tags":["openapi"],"summary":"ui openapi","description":"Redirect to Swagger UI","operationId":"openapi#ui","responses":{"301":{"description":"Moved Permanently response.","headers":{"Location":{"description":"Redirect location","type":"string"}}}},"schemes":["https"]}},"/tenants":{"get":{"tags":["tenant"],"summary":"list tenant","description":"List all tenants","operationId":"tenant#list","responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/Tenant"}}}},"schemes":["https"]},"post":{"tags":["tenant"],"summary":"create tenant","description":"Create a new tenant","operationId":"tenant#create","produces":["application/json"],"parameters":[{"name":"CreateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/TenantCreateRequestBody","required":["name","slug","email"]}}],"responses":{"201":{"description":"Created response.","schema":{"$ref":"#/definitions/Tenant"}}},"schemes":["https"]}},"/tenants/{id}":{"get":{"tags":["tenant"],"summary":"get tenant","description":"Get tenant by ID","operationId":"tenant#get","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Tenant ID","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/Tenant"}}},"schemes":["https"]},"put":{"tags":["tenant"],"summary":"update tenant","description":"Update tenant","operationId":"tenant#update","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Tenant ID","required":true,"type":"string","format":"uuid"},{"name":"UpdateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/TenantUpdateRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/Tenant"}}},"schemes":["https"]},"delete":{"tags":["tenant"],"summary":"delete tenant","description":"Delete tenant","operationId":"tenant#delete","parameters":[{"name":"id","in":"path","description":"Tenant ID","required":true,"type":"string","format":"uuid"}],"responses":{"204":{"description":"No Content response."}},"schemes":["https"]}}},"definitions":{"OpenapiSpecInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"spec_internal_error_response_body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"OpenapiSpecResponseBody":{"title":"OpenapiSpecResponseBody","type":"object","properties":{"spec":{"description":"OpenAPI specification","example":{"info":{"title":"Enterprise AWO ERP System API","version":"1.0.0"},"openapi":"3.0.3"}}},"example":{"spec":{"info":{"title":"Enterprise AWO ERP System API","version":"1.0.0"},"openapi":"3.0.3"}},"required":["spec"]},"Tenant":{"title":"Mediatype identifier: application/vnd.tenant; view=default","type":"object","properties":{"created_at":{"type":"string","description":"Creation timestamp","example":"2023-12-07T10:30:00Z","format":"date-time"},"currency_code":{"type":"string","description":"ISO 4217 currency code for UI display","default":"USD","example":"USD","pattern":"^[A-Z]{3}$"},"email":{"type":"string","description":"Used for administrative communications","example":"admin@acme.com","format":"email"},"id":{"type":"string","description":"Unique tenant identifier","example":"550e8400-e29b-41d4-a716-446655440000","format":"uuid"},"name":{"type":"string","description":"Display name shown in UI","example":"ACME Corporation","minLength":2,"maxLength":100},"plan":{"type":"string","description":"Determines feature access and limits","default":"FREE","example":"PROFESSIONAL","enum":["FREE","STARTER","PROFESSIONAL","ENTERPRISE","CUSTOM"]},"slug":{"type":"string","description":"Used in URLs and subdomain routing. Must be globally unique.","example":"acme-corp","pattern":"^[a-z0-9][a-z0-9-]{1,48}[a-z0-9]$","minLength":3,"maxLength":50},"status":{"type":"string","description":"Controls tenant access and billing","default":"TRIAL","example":"ACTIVE","enum":["ACTIVE","SUSPENDED","TRIAL","ARCHIVED","PENDING_SETUP"]},"timezone":{"type":"string","description":"IANA timezone identifier","default":"UTC","example":"America/New_York","pattern":"^[A-Za-z_]+/[A-Za-z_]+$"},"updated_at":{"type":"string","description":"Last update timestamp","example":"2023-12-07T15:45:00Z","format":"date-time"}},"description":"Business tenant with complete multi-tenancy isolation and configuration support (default view)","example":{"created_at":"2023-12-07T10:30:00Z","currency_code":"USD","email":"admin@acme.com","id":"550e8400-e29b-41d4-a716-446655440000","name":"ACME Corporation","plan":"PROFESSIONAL","slug":"acme-corp","status":"ACTIVE","timezone":"America/New_York","updated_at":"2023-12-07T15:45:00Z"},"required":["id","slug","name","email","status","timezone","currency_code","created_at"]},"TenantCreateRequestBody":{"title":"TenantCreateRequestBody","type":"object","properties":{"email":{"type":"string","description":"Contact email","example":"Nobis consequatur enim eius esse."},"name":{"type":"string","description":"Tenant name","example":"Omnis occaecati quasi veniam."},"slug":{"type":"string","description":"Tenant slug","example":"Dolorem est repellat sint error vitae mollitia."}},"example":{"email":"Autem ea sequi qui numquam id.","name":"Qui cumque debitis rerum consequatur et.","slug":"Harum ut eaque ad ratione."},"required":["name","slug","email"]},"TenantUpdateRequestBody":{"title":"TenantUpdateRequestBody","type":"object","properties":{"email":{"type":"string","description":"Contact email","example":"Omnis at voluptas temporibus."},"name":{"type":"string","description":"Tenant name","example":"Voluptatem voluptatem asperiores error placeat est."}},"example":{"email":"Voluptates quasi enim nihil at sint dolores.","name":"Ut quia corporis dolorem odio error."}}}}